Chore: Refactor events folder for clarification
